+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Add support for cover lock
 - Add more information to sensors and binary sensors in the HA integration
+- Store last Telegram and decoded value in RemoteValue
 - Drop support for python 3.7
 
 ### Internals